add additional documentation
This documents the remaining entities in the code, and sets things up
with `yard` to make it able to generate locally. It has not been able to
generate on https://rdoc.info itself. Perhaps the error is with
`org-mode` files (which were problematic before installing `org-ruby`)
or in th prior state of documentation. Either way, these changes present
a pristine repository that might help us with our rdoc issue.

No runtime changes were made.