add into_result on Array, add Result#match
This adds an `#into_result` on `Array`. This way an `Array` of `Results`
can easily be coerced into a `Result` of an `Array`. If any of the
members are an `Err` then the returned `Result` is an `Err` of all of
the `Err` values as an `Array`. Otherwise it returns an `Ok` of all of
the `Ok` values, as an `Array`. To accommodate this, `Err` has been
relaxed and now accepts any value, not just `Exceptions`.

To help build this functionality, `#match` had to be added to `Result`.

Earmarked in this change is:
1. Some methods and tests were rearranged for lexicographical order.
2. Parts of the `README` are now executable!