add Either
This adds the `Either` type, which has `Left` and `Right` variants as opposed to
Result's `Ok` and `Err`.  This is an important distinction because there is no
bias for one side over the other.

There's some additional documentation we could put into the README, and it's not
on feature parity with Rust's `Either`.  It should also be noted that Rust's
`Either` is not part of the core but instead a library.  Some of the function
names are kind of weird (like `left_and_then` instead of `and_then_left` that
would match everything else).  It's unknown why that is, and we might make an
alias for it.  We might also introduce `and_then`, `map`, and other common
interfaces that assume `Right`, which I'm pretty sure how Haskell does it.
Aliasing methods in Ruby is cheap/easy, so this shouldn't be a problem.