add in support for coercing Futures to Results
This adds an optional file (`monad-oxide/concurrent-promises`) which is designed
to work with `concurrent-ruby`'s `Future`.  It adds `#into_result` onto `Future`
and that method returns a `MonadOxide::Result`, whose specific type depends upon
whether or not the `Future` was successful.

As part of testing this, I found out that RubyGems basically enforces that `lib`
only ever contains the "entry point" of the gem.  All other supporting files
_must_ be located under the name of that gem.  While things can technically work
without adhering to that standard, it causes pollution in the `$LOAD_PATH`.  So
the files have been moved to accommodate this.

Additionally the `frozen_string_literal` stamp has been added to all files -
something all Ruby files should have by default.